+++ src/components/Toast/style.js
@@ -1,21 +1,24 @@
-import Styled, { keyframes } from 'styled-components';
+import Styled from 'styled-components';
+import theme from '../../styles/theme';
 
 export const Container = Styled.div`
   display: flex;
   justify-content: space-between;
   align-items: center;
-  background: ${props => props.color || '#4fa444'};
+  background: ${props => props.color || theme.successColor};
   padding: 0px 20px;
   width: 100%;
   height: 40px;
   position: absolute;
   top: 60px;
   z-index: 2;
-  transform: translateY(${ props => props.active ? '0px' : '-100px' });
+  transform: translateY(${props => (props.active ? '0px' : '-100px')});
   transition: all 0.3s ease;
 `;
 
-export const IconBox = Styled.div``;
+export const IconBox = Styled.div`
+  display: flex;
+`;
 
 export const MessageBox = Styled.span`
   font-size: 14px;