+++ src/index.js
@@ -1,8 +1,9 @@
-export { default as Modal } from "./components/Modal";
-export { default as Checkbox } from "./components/Checkbox";
-export { default as Tooltip } from "./components/Tooltip";
-export { default as SentFeedback } from "./components/SentFeedback";
-export { default as Table } from "./components/Table";
-export { default as Balance } from "./components/Balance";
-export { default as Button } from "./components/Button";
-export { default as Label } from "./components/Label";
+export { default as Modal } from './components/Modal';
+export { default as Checkbox } from './components/Checkbox';
+export { default as Tooltip } from './components/Tooltip';
+export { default as SentFeedback } from './components/SentFeedback';
+export { default as Table } from './components/Table';
+export { default as Balance } from './components/Balance';
+export { default as Button } from './components/Button';
+export { default as Label } from './components/Label';
+export { default as Toast } from './components/Toast';