+++ stories/button.story.js
@@ -1,20 +1,20 @@
 /* eslint-disable import/no-extraneous-dependencies */
-import React from "react";
-import { storiesOf } from "@storybook/react";
-import Button from "../src/components/Button";
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+import Button from '../src/components/Button';
 
 const fileCharacteristic = {
-  id: "testezin",
+  id: 'testezin',
   multiple: false,
-  name: "teste",
+  name: 'teste',
   onChange: e => {
-    console.log("FILE => ", e.target.value);
-  }
+    console.log('FILE => ', e.target.value);
+  },
 };
 
-storiesOf("Buttons")
-  .add("Main", () => <Button>Primary</Button>)
-  .add("Secondary", () => <Button isSecondary>Secondary</Button>)
-  .add("Main to send File", () => (
+storiesOf('Buttons')
+  .add('Main', () => <Button>Primary</Button>)
+  .add('Secondary', () => <Button isSecondary>Secondary</Button>)
+  .add('Main to send File', () => (
     <Button fileCharacteristic={fileCharacteristic}>Get File</Button>
   ));